% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aneeldata}
\alias{aneeldata}
\title{ANEEL database on efficiency scores.}
\format{
An object of class \code{list} of length 2.
}
\usage{
data("aneeldata", package = "gbdcd")
}
\description{
List containing the eficiency for each company and a matrix indicating the neighbors.
}
\keyword{datasets}
