% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aneelshape}
\alias{aneelshape}
\title{ANEEL database on efficiency scores, environmental and geographical data.}
\format{An object of class \code{SpatialPolygonsDataFrame} with 57 rows and 151 columns.}
\usage{
data("aneelshape", package = "gbdcd")
}
\description{
ANEEL database on efficiency scores, environmental and geographical data.
}
\keyword{datasets}
