% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{gaussianBDCD}
\alias{gaussianBDCD}
\title{Gaussian BDCD}
\usage{
gaussianBDCD(y, viz, c = 0.35, n_iterations = 1e+06, burn_in = 5e+05,
  mu0 = 0, sigma0 = sqrt(2))
}
\arguments{
\item{parameter.name}{}
}
\value{
\code{data.frame}
}
\description{
Description
}
\examples{

}
\author{
leandromineti@nexer.com.br
}
