% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{gbdcd_regression}
\alias{gbdcd_regression}
\title{Gaussian BDCD regression}
\usage{
gbdcd_regression(
  y,
  x,
  viz,
  n_iterations = 1e+05,
  burn_in = 50000,
  c = 0.35,
  prior_coeffs_mu = rep(0, 2),
  prior_sigma = sqrt(2),
  lambda = 0.001,
  plot = F
)
}
\arguments{
\item{y}{a vector specifying the target variable for each of the nodes.}

\item{x}{a matrix with the independent variables.}

\item{viz}{a matrix defining the graph neighbors.}

\item{n_iterations}{number of iterations.}

\item{burn_in}{number of discarded iterations.}

\item{c}{parameter indicating the \emph{a priori} number of clusters.}

\item{prior_coeffs_mu}{\emph{a priori} mean for regresion coefficients.}

\item{prior_sigma}{\emph{a priori} standard deviation.}

\item{lambda}{regularization parameter.}

\item{plot}{plot the results. Default = FALSE.}
}
\value{
a \code{list} of seven objects:
\itemize{
  \item mean.info: \emph{a posteriori} means and credible interval.
  \item cluster.info: hierarchical clustering return object.
  \item matConnections: frequency matrix indicating how many times each
  pair of nodes was in the same cluster.
  \item k.MCMC: a vector indicating the number of clusters in each iteration.
  \item mean_y: target variable mean.
  \item sd_y: target variable standard deviation.
  \item vec.centers: center configuration for each iteration.
}
}
\description{
Implementation of the Bayesian Detection of Clusters and Discontinuities
}
\seealso{
Other gbdcd: 
\code{\link{gbdcdGroups}()},
\code{\link{gbdcd}()}
}
\author{
leandromineti@gmail.com
}
\concept{gbdcd}
