% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RcppFreqMatrix}
\alias{RcppFreqMatrix}
\title{Frequency matrix generation}
\usage{
RcppFreqMatrix(partitions)
}
\arguments{
\item{partitions}{A vector indicating the cluster center for each graph node.}
}
\value{
a matrix.
}
\description{
Generate frequency matrix
}
\details{
Frequency matrix
}
\author{
Leandro Mineti
}
