% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{gaussianBDCD}
\alias{gaussianBDCD}
\title{Gaussian BDCD}
\usage{
gaussianBDCD(y, viz, c = 0.35, n_iterations = 1e+06, burn_in = 5e+05,
  mu0 = 0, sigma0 = sqrt(2))
}
\arguments{
\item{y}{a vector of indices.}

\item{viz}{a matrix defining the neighbors.}

\item{c}{parameter indicating the a priori number of clusters.}

\item{n_iterations}{number of iterations.}

\item{burn_in}{number of discarded iterations.}

\item{mu0}{priori.}

\item{sigma0}{priori.}
}
\value{
\code{list}
}
\description{
Implementation of the Bayesian Detection of Clusters and Discontinuities
}
\examples{

}
\author{
leandromineti@gmail.com
}
