% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussian.R
\name{gbdcdGroups}
\alias{gbdcdGroups}
\title{GBDCD groups}
\usage{
gbdcdGroups(out, k = 2)
}
\arguments{
\item{out}{output from gbdcd.}

\item{k}{number of clusters}
}
\value{
a \code{list} of seven objects:
\itemize{
  \item mydata: 
  \item fit: 
  \item groups: 
}
}
\description{
Implement Adapted Ng-Jordan-Weiss clustering on gbdcd results
}
\examples{
library(gbdcd)

data("aneeldata", package = "gbdcd")
data("aneelshape", package = "gbdcd")

target_variable <- aneelshape$z_Precipitation
neighbors <- aneeldata$connections

out <- gaussianBDCD(y = target_variable, 
                   neigh = neighbors, 
                   c = 0.35, 
                   n_iterations = 100000, 
                   burn_in = 50000, 
                   mu0 = 0, 
                   sigma0 = sqrt(2))
                   
clusters <-gbdcdGroups(out, k = 3)

}
\seealso{
Other gbdcd: \code{\link{gaussianBDCD}}
}
\author{
leandromineti@gmail.com
}
